Move Encoder quality API to VideoEncoderSettings.
Some other minor nits and adjustments to the API logic.

PiperOrigin-RevId: 459490431
(cherry picked from commit 91f1777741149ecf67759da8eb68828e1d92624b)

+++ libraries/transformer/src/main/java/androidx/media3/transformer/EncoderBitrateProvider.java
@@ -26,11 +26,11 @@ public interface EncoderBitrateProvider {
   /**
    * Returns a recommended bitrate that the encoder should target.
    *
-   * @param encoderName The name of the encoder, see {@link MediaCodecInfo#getName()}
+   * @param encoderName The name of the encoder, see {@link MediaCodecInfo#getName()}.
    * @param width The output width of the video after encoding.
    * @param height The output height of the video after encoding.
    * @param frameRate The expected output frame rate of the video after encoding.
    * @return The bitrate the encoder should target.
    */
-  int getBitrate(String encoderName, int width, int height, int frameRate);
+  int getBitrate(String encoderName, int width, int height, float frameRate);
 }
